add cors

+++ init-db/01-init.sql
@@ -3,6 +3,6 @@ CREATE TABLE IF NOT EXISTS clientes (
     nome VARCHAR(255) NOT NULL,
     email VARCHAR(255) UNIQUE NOT NULL,
     telefone VARCHAR(20),
-    coordenada_x INTEGER,
-    coordenada_y INTEGER
+    coordenada_x bigint,
+    coordenada_y bigint
 );
